echo example: make actors less assuming what the next message type is
I.e before this change, Input, Damper actors assumed they are fed into Mixer,
but Delay assumed it is fed into something that expects naked `Chunk`.
This made no sense, and made actors dependent on how they are wired in the
system, which is an antipattern.

This is an artificial change that demonstrates problems outlined in
https://github.com/tonarino/actor/issues/11#issuecomment-783110918