Implement message receive timeout, the Actor::deadline_passed() callback
This is a building block for timers, but a full-fledged timer is not actually
implemented here. A simplified version of such in presented as an example.

It is made possible by ability to set a timeout for crossbeam-channel select!
macro/call.

In theory, the deadline can be used beyond timers, for example for "yielding"
from some long-running computation to check for new messages, including
control messages. But crossbeam-channel select() says
> If multiple operations are ready at the same time, a random one among them is selected.

..so this is only a "stochastic yield".

v3: go back to setting the deadline through Context (but without setters this
    time), preserve deadline when a message is received, add test, provide 2
    separate examples (delay actor now with per-message recipient and simple
    timer).