Context: change myself field from Addr<A> to Recipient<M>
This makes whole Context parametrised by M (message), rather than A: Actor.
As a result, Context is more flexible to pass on, which facilitates creation
of Actor wrappers.

This changes the signature of the Actor trait methods, which breaks the API,
though required edits should be mostly mechanical. The field `myself` was
arguably used most for sending messages to oneself, which already went
through a `Deref` to `Recipient`.

One piece of functionality was lost on Context: the `Addr::recipient()`
method which can produce a "converting" recipient.
`recipient()` can be moved from `Addr` to `Recipient` to preserve it.