Add Box<dyn fmt::Debug> to SendError, require Actor::Message: Debug
The problem: does not compile, because popular error libraries like `anyhow`
require the error to be `Sync`. Because we erase the `M` type from SendError,
we either require `M` to be `Sync` or we're not `Sync` at all.

(crosbeam::channel's TrySendError<M> doesn't suffer from such a problem,
it is Send when M is send and vice-versa).