RFC: Include intended recipient name in SendError (#59)
* Add test for SendError

We're going to touch it, the tests will demonstrate the changes.

* Include intended recipient name in `SendError`

This aids with debugging send errors. Thanks to type decoupling made by
`Recipient`, senders may not even know the actual receiving actor, so
senders cannot replicate this behaviour fully on their own.

Should help with debugging downstream issues like tonarino/portal#1153.

Caveats:
- Increases the size of `Recipient`, `Addr`, `SendError` by two pointers (one &str).
- Introduces API-breaking changes to `SendError` and `SendResultExt::on_full()`.

* Add comments/docstrings (per review comments)

* Convert SendError fields to pub

Rationale: if we ever want to contain original message `M` as a payload
in the error, it would be more handy if it is pub so that users can move
the message out of the error (without introducing special methods for it).