Really make the errors() test deterministic
I claimed this in the previous PR, but it wasn't true: if the actor spawned quick enough the test
would fail. What _was_ true is that I've ran the testsuite hundreds of times in a loop, and it
didn't fail. Some strange coincidence?

Make it truly deterministic by not spawning the actor at all (test using a naked `Addr`).