Move `.recipient()` from `Addr` to `Recipient`, Capacity fields public (#65)
* Really make the errors() test deterministic

I claimed this in the previous PR, but it wasn't true: if the actor spawned quick enough the test
would fail. What _was_ true is that I've ran the testsuite hundreds of times in a loop, and it
didn't fail. Some strange coincidence?

Make it truly deterministic by not spawning the actor at all (test using a naked `Addr`).

* Move `.recipient()` from `Addr` to `Recipient`

This makes the API more flexible as we now can call `.recipient()` multiple times in a row on the
same value, allowing one more M -> N message conversion in each step.

Previously, we thought it is too wild to create "arbitrarily long conversion chain", so we've put
`.recipient()` on `Addr`, only allowing one such conversion.

But with the advent of actor wrappers like `Timed<A>`, this first conversion is "consumed" by the
wrapper itself, limiting their users.

The method name `Recipient::recipient()` seems a bit weird, but there's a high value in keeping the
current usage working without any changes (because Addr derefs to Recipient), and no obviously
better name came into my mind.

* Make Capacity fields (high, normal) public

This is an API omission discovered when we tried to set both normal and priority inbox sizes to
different and non-default values.

I argue that `Capacity { normal: Some(3), high: Some(9) }` is more explicit than
`Capacity::new(3, 9)`, hence making the fields pub instead of adding a constructor.

* Add example of .recipient() chaining to delay_actor example