Reduce duplication between `Context` and `TimedContext` using `BareContext`
Motivation for this:
1. (small) reduce logical duplication
2. (big) pave the way for async actors, that will also want to use `BareContext` as their context.

This is de-facto semver compatible, thanks to the Deref impls. Only subtle difference is that it is no longer possible to obtain `&mut SystemHandle` nor `&mut Recipient<MyMessage>` from `&mut Context`. You can only get `&` reference (or the owned values by cloning them). But it doesn't matter, there are no public `&mut` methods on these, so `&` is always sufficient.

I've tried this with our portal code, and indeed the only change I needed to do was `&mut` -> `&` at 2 places. This change can be done with the released actor 0.11 already, the `mut` is strictly unnecessary there.